Update ensureIndexes signature to return a Future instead of Unit
The ensureIndexes method on the Indexes trait was returning Unit, even
though the underlying calls to ReactiveMongo will almost always be
returning a Future.  Updated this to allow the Future to be returned.

This allows test code to await the index building repsonse, and gives
production code the option to handle errors that may occur during the
creation of the index.