PE-1002 - Simple reactive mongo tests are flaky
The former test relied on  a race condition in index creation.
When reactiveMongo is creating an index it first checks if the index had already been created. If so, it returns false but doesn't throw any exception.

However, if two indexes with the same name are created simultaneously, the early test doesn't prevent the index creation. One of the indexes is eventually created and reactivemongo eventually throws an exception when it tries to create the second index.

Since the goal of the test was to detect exceptions on index creation, we now force an exception by creating a unique index on a collection with duplicate entries.