INFRA-5071: list indexes for non-existant collection throws
In order to fix https://github.com/ReactiveMongo/ReactiveMongo/issues/700,
the reactivemongo maintainers changed the serialiser they use.
Now the error returned whenever there is no collection detected
_isn't_ a WriterResult, so doesn't get caught and converted
into a Future.successful(List.empty). Instead, the error gets
propogated out to the user.

Since listing the indexes on a collection that doesn't exist is,
at best, a testing activity (and we've not seen it in our review
of production code), I'm going to put this on the list of 'upgrade
gotchas' and update the tests to account for the difference.